+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/exception/NoHandlerAdapterFoundException.java
@@ -1,0 +1,7 @@
+package com.interface21.webmvc.servlet.mvc.tobe.exception;
+
+public class NoHandlerAdapterFoundException extends RuntimeException {
+    public NoHandlerAdapterFoundException(String message) {
+        super(message);
+    }
+}
