+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/exception/NoHandlerFoundException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.exception;
 
 public class NoHandlerFoundException extends RuntimeException {
     public NoHandlerFoundException(String message) {
