+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/exception/NoHandlerMappingFoundException.java
@@ -1,0 +1,7 @@
+package com.interface21.webmvc.servlet.mvc.tobe.exception;
+
+public class NoHandlerMappingFoundException extends RuntimeException {
+    public NoHandlerMappingFoundException(String message) {
+        super(message);
+    }
+}
